Update README
Why
---
The current instructions require a developer to follow multiple READMEs
in multiple repositories (not explicitly referenced here) in order to
use this setup.  I want a developer to be able to follow a single set of
instructions and get from zero to running Hubs in minimal time and
without pain.

What
----
* Add hosts file entries to Initial Setup
* Add Self-Signed Certificates section
* Change the Mutagen Compose link to point to the System Requirements
  section which outlines its dependencies