Fix scripts
Why
---

On a fresh install the initialization script may fail with a read-only
file system error.  This is because `readlink` is being invoked on the
services directory before it exists.

When the path contains a space all of the scripts break.  This is
because the commands treat the path as multiple arguments.

What
----

* Apply `readlink` to a directory that is sure to exist
* Quote paths