Update the timeout troubleshooting section to be less dangerous
The current procedure for solving timeouts when downloading dependencies is dangerous because it embeds a potentially infinite (if some other non-timeout related error is encountered) while loop in a file that is copied onto the docker image that will persist, essentially hidden, until the conditional-npm-ci file is updated and your docker images are regenerated (probably by bin/reset) because any changes to the conditional-npm-ci file on your file system won't affect the conditional-npm-ci file on the docker image (which is the one that is actually being run by the scripts).

Instead, we now instruct the user to add the command to the bin/init script which is run directly, not through docker, and so will execute any changes immediately and predictably.