Fix errors when switching branches in service repositories
What:
Changes docker-compose.yml to make Mutagen use `one-way-replica` as the sync mode.

Why:
For some reason, using `two-way-resolved` would sometimes not accept the changes from the branch switch and actually push some of the old code from the previous branch in the Docker container back to your filesystem, this would result in both the Docker container and your file system being in a halfway state between branches and cause errors with the affected service.  Changing to `one-way-replica` forces the Docker container to always mirror your file system and prevents this problem.

Note:
This was mainly noticed when switching between the master branch and the addons branch in the `hubs` repository (the branches have different dependencies, which could possibly be a contributing factor).